fix: repair cursor delta landing gates
