+++ ui/src/pages/chat/chat-agent-id.ts
@@ -1,0 +1,15 @@
+import { scopedAgentParamsForSession } from "../../lib/sessions/index.ts";
+import {
+  normalizeAgentId,
+  parseAgentSessionKey,
+  resolveUiSelectedGlobalAgentId,
+} from "../../lib/sessions/session-key.ts";
+import type { ChatState } from "./chat-state-contract.ts";
+
+export function resolveChatAgentId(state: ChatState) {
+  return normalizeAgentId(
+    parseAgentSessionKey(state.sessionKey)?.agentId ??
+      scopedAgentParamsForSession(state, state.sessionKey).agentId ??
+      resolveUiSelectedGlobalAgentId(state),
+  );
+}
